Arbitrary resource types - base implementation
Replace hardcoded instantiation of JarResource in RepoIndex by a
loop that consults all available recognizers, similar to how the
analyzers are consulted.

Added the Default Recognizer that recognizes bundles and jars
and returns them as a Jar Resource with the appropriate mime
type. Or, otherwise returns a File Resource with mime type
unknown.

Added the Default Analayzer that tries to determine generic
identity capabilities based on filename for File Resources
(as a fallback).

Modified registration of existing (bundle) analyzers to not
consider mimetypes they can not handle.